Move IntegrationTestFrameworkHelpers to Shared
