--- tests/Agent/IntegrationTests/ApplicationHelperLibraries/IntegrationTestFrameworkHelpers/IntegrationTestFrameworkHelpers.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-</Project>
