+++ tests/Agent/IntegrationTests/IntegrationTests.sln
@@ -115,8 +115,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ApplicationHelpers", "Appli
 		ApplicationHelperLibraries\readme.md = ApplicationHelperLibraries\readme.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntegrationTestFrameworkHelpers", "ApplicationHelperLibraries\IntegrationTestFrameworkHelpers\IntegrationTestFrameworkHelpers.csproj", "{2D9C79C0-0C7B-4816-BE7E-1DB9EBF59667}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleOtherTransactionWrapperApplication", "Applications\ConsoleOtherTransactionWrapperApplication\ConsoleOtherTransactionWrapperApplication.csproj", "{5EF05574-5A01-404B-92B0-728A46577FDC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleMultiFunctionApplicationFW", "SharedApplications\ConsoleMultiFunctionApplicationFW\ConsoleMultiFunctionApplicationFW.csproj", "{C37AA470-FFF4-4A61-B240-F858CF1357C5}"
@@ -325,10 +323,6 @@ Global
 		{C3D78B19-80F7-4210-B848-A5AA2BF5D13F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C3D78B19-80F7-4210-B848-A5AA2BF5D13F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C3D78B19-80F7-4210-B848-A5AA2BF5D13F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2D9C79C0-0C7B-4816-BE7E-1DB9EBF59667}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2D9C79C0-0C7B-4816-BE7E-1DB9EBF59667}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2D9C79C0-0C7B-4816-BE7E-1DB9EBF59667}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2D9C79C0-0C7B-4816-BE7E-1DB9EBF59667}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5EF05574-5A01-404B-92B0-728A46577FDC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5EF05574-5A01-404B-92B0-728A46577FDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5EF05574-5A01-404B-92B0-728A46577FDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -412,7 +406,6 @@ Global
 		{83FA6862-73F2-4305-9B7C-2DF849D80C36} = {F0F6F2CE-8AE8-49E1-8EE9-A44B451EFC29}
 		{C3D78B19-80F7-4210-B848-A5AA2BF5D13F} = {F0F6F2CE-8AE8-49E1-8EE9-A44B451EFC29}
 		{CB5B7DCD-F949-4435-BC39-C34F4F3D2009} = {F0F6F2CE-8AE8-49E1-8EE9-A44B451EFC29}
-		{2D9C79C0-0C7B-4816-BE7E-1DB9EBF59667} = {CB5B7DCD-F949-4435-BC39-C34F4F3D2009}
 		{5EF05574-5A01-404B-92B0-728A46577FDC} = {F0F6F2CE-8AE8-49E1-8EE9-A44B451EFC29}
 		{C37AA470-FFF4-4A61-B240-F858CF1357C5} = {AD5B3EF6-1222-46B1-8FD4-6084FF75FA30}
 		{4213F281-732C-4447-B347-1C46E3609986} = {AD5B3EF6-1222-46B1-8FD4-6084FF75FA30}
